Workaround #30 "'str' does not support the buffer interface" (libxml2)
python3-libxml2 breaks xml.sax.  Test for this specific bug.

Code assumes the systems affected by the bug, have an alternative
sax provider which will be preferred.  Otherwise, we'll still die.
All reports (see the issue) are that this workaround is effective.
We don't have to worry too much about the future either, as the bug
is fixed upstream.